refactor(website): move the browser-safe helpers into lib/utils
AGENTS.md puts browser-safe pure helpers in lib/utils/ and keeps lib/*.ts for
app-wide values. highlight.ts, frontmatter.ts and faq.ts were sitting loose at
the top level, which is part of why lib/ read as a junk drawer: four different
kinds of thing at one level with nothing separating them.

lib/ now holds the design system (design, brand, brand-tokens), shared SSR
composition (site-footer, docs-shell), config and content (links, samples),
one server-only module, and lib/utils/ for the three helpers.

The alias rewrite missed two classes of reference that only a full test run
surfaced: a test that READS a helper by literal path rather than importing it,
and doc comments naming the old location. Both are fixed. 35 tests pass.
