+++ website/modules/blog/queries/get-post.server.ts
@@ -3,7 +3,7 @@
 import { readFile } from 'node:fs/promises';
 import { join, dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { parseFrontmatter } from '#lib/frontmatter.ts';
+import { parseFrontmatter } from '#lib/utils/frontmatter.ts';
 import type { PostWithBody } from '#modules/blog/types.ts';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
