+++ website/modules/blog/queries/list-posts.server.ts
@@ -3,7 +3,7 @@
 import { readdir, readFile } from 'node:fs/promises';
 import { join, dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { parseFrontmatter } from '#lib/frontmatter.ts';
+import { parseFrontmatter } from '#lib/utils/frontmatter.ts';
 import type { Post } from '#modules/blog/types.ts';
 
 // website/modules/blog/queries/list-posts.server.ts is 4 levels deep
